+++ src/exceptions/UnsupportedLabelFormatException.php
@@ -1,0 +1,7 @@
+<?php
+namespace verbb\shippy\exceptions;
+
+class UnsupportedLabelFormatException extends InvalidRequestException
+{
+
+}